\name{Data}
\docType{data}
\alias{Data}

\title{Customer journeys data.}

\description{
 Example dataset. 
}
\usage{data(PathData)}
\format{
  \code{Data} is a data.frame with 10.000 rows and 4 columns: "path" containing customer paths, "total_conversions" containing total number of conversions, "total_conversion_value" containing total conversion value and "total_null" containing total number of paths that do not lead to conversion.
}

\keyword{customer journey dataset}
\keyword{customer path data}
\keyword{dataset}

