\name{Dy}
\docType{data}
\alias{Dy}
\title{Customer journeys data}
\description{
 Customer path data with conversions and conversion value. 
}
\usage{data(PathData)}
\format{
  \code{Dy} is a data.frame with 10.000 rows and 3 columns: "channel_path" containing customer paths, "total_conversions" containing total number of conversions and "total_conversion_value" containing total conversion value.
}
\keyword{online attribution}
\keyword{channel attribution}
\keyword{multi channel funnel}
\keyword{customer journey}
\keyword{customer path}
\keyword{dataset}

